0.2.0 | Change name to todo-note instead of todo-des
Todo.txt Task **does** have a description so `des` here causes confusion
for the users.

A `note` for a Task is more understandable and does not introduce
conflicting concepts.

*new feature:*

- `add` allows to use an external editor (by omitting the `-n`). The
note's content is no longer a positional argument in the cli. You have
to use `-n` to supply the content (inspired by `git commit`).